% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowsColsQ.R, R/old_fns.R
\name{build_unpivot_control}
\alias{build_unpivot_control}
\alias{buildUnPivotControlTable}
\title{Build a moveValuesToColumns*() control table that specifies a un-pivot (or "shred").}
\usage{
build_unpivot_control(nameForNewKeyColumn, nameForNewValueColumn,
  columnsToTakeFrom, ...)

buildUnPivotControlTable(nameForNewKeyColumn, nameForNewValueColumn,
  columnsToTakeFrom, ...)
}
\arguments{
\item{nameForNewKeyColumn}{character name of column to write new keys in.}

\item{nameForNewValueColumn}{character name of column to write new values in.}

\item{columnsToTakeFrom}{character array names of columns to take values from.}

\item{...}{not used, force later args to be by name}
}
\value{
control table
}
\description{
Some discussion and examples can be found here:
\url{https://winvector.github.io/FluidData/FluidData.html} and
here \url{https://github.com/WinVector/cdata}.

buildUnPivotControlTable deprecated, functionality moved to: build_unpivot_control
}
\examples{

build_unpivot_control("measurmentType", "measurmentValue", c("c1", "c2"))

}
\seealso{
\code{\link{rowrecs_to_blocks_q}}, \code{\link{rowrecs_to_blocks}}
}
