% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cdata-package}
\alias{cdata}
\alias{cdata-package}
\title{\code{cdata}: Fluid Data Transformations.}
\description{
Supplies implementations of higher order "fluid data" transforms.  These
transforms move data between rows and columns, are controlled by a graphical
transformation specification, and have pivot and un-pivot as special cases.
Large scale implementation is based on 'rquery', so should be usable on
'SQL' compliant data sources (include large systems such as 'PostgreSQL' and
'Spark').
This package introduces the idea of  control table specification of data transforms (later aslo adapted from 'cdata' by 'tidyr').
A theory of fluid data transforms can be found in the following articles:
\url{https://winvector.github.io/FluidData/FluidDataReshapingWithCdata.html},
\url{https://github.com/WinVector/cdata} and \url{https://winvector.github.io/FluidData/FluidData.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WinVector/cdata/}
  \item \url{https://winvector.github.io/cdata/}
  \item Report bugs at \url{https://github.com/WinVector/cdata/issues}
}

}
\author{
\strong{Maintainer}: John Mount \email{jmount@win-vector.com}

Authors:
\itemize{
  \item Nina Zumel \email{nzumel@win-vector.com}
}

Other contributors:
\itemize{
  \item  Win-Vector LLC [copyright holder]
}

}
