% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cdata}
\alias{cdata}
\alias{cdata-package}
\title{cdata: wrappers for \code{tidyr::gather()} and \code{tidyr::spread()}}
\description{
cdata supplies deliberately verbose wrappers for \code{tidyr::gather()} and \code{tidyr::spread()} and an explanatory vignette.
}
