% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf}
\alias{cf}
\title{Build character data.frame}
\usage{
cf(...)
}
\arguments{
\item{...}{character values, + denotes end of row, first row is column names.}
}
\value{
character data.frame
}
\description{
A confineint way to build a character data.frame.  Position of
first "+" determines number of columns (all other "+"'s are aliases for ",").
}
\examples{

cf("measure",                    "training", "validation" +
   "minus binary cross entropy", "loss",     "val_loss"   +
   "accuracy",                   "acc",      "val_acc"    )

}
