% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{char_frame}
\alias{char_frame}
\title{Build a (non-empty) character data.frame.}
\usage{
char_frame(..., cf_eval_environment = parent.frame())
}
\arguments{
\item{...}{cell names, first infix operator denotes end of header row of column names.}

\item{cf_eval_environment}{environment to evaluate names in.}
}
\value{
character data.frame
}
\description{
A convenient way to build a character data.frame in legible transposed form.  Position of
first "+" (or other infix operator) determines number of columns
(all other infix operators are aliases for ",").
Names are de-referenced.
}
\examples{

tc_name <- "training"
x <- char_frame(
   "measure",                   tc_name, "validation" /
   "minus binary cross entropy", "loss", "val_loss"   /
   "accuracy",                    "acc", "val_acc"    )
print(x)
cat(draw_frame(x))

}
\seealso{
\code{\link{build_frame}},  \code{\link{qchar_frame}}, \code{\link{build_frame}}
}
