% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cols_form_unique_keys.R
\name{check_cols_form_unique_keys}
\alias{check_cols_form_unique_keys}
\title{Check columns form unique keys}
\usage{
check_cols_form_unique_keys(data, keyColNames)
}
\arguments{
\item{data}{data.frame to check}

\item{keyColNames}{character, names of columns to consider as keys}
}
\value{
logical TRUE if rows are uniquely keyed by named columns
}
\description{
Check columns form unique keys
}
\keyword{internal}
