% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdata_yaml.R
\name{convert_cdata_spec_to_yaml}
\alias{convert_cdata_spec_to_yaml}
\title{Convert a layout_specification, blocks_to_rowrecs_spec, or rowrecs_to_blocks_spec to a simple object.}
\usage{
convert_cdata_spec_to_yaml(spec)
}
\arguments{
\item{spec}{a layout_specification, blocks_to_rowrecs_spec, or rowrecs_to_blocks_spec}
}
\value{
a simple object suitable for YAML serialization
}
\description{
Convert a layout_specification, blocks_to_rowrecs_spec, or rowrecs_to_blocks_spec to a simple object.
}
