% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdata_yaml.R
\name{convert_yaml_to_cdata_spec}
\alias{convert_yaml_to_cdata_spec}
\title{Read a cdata record transform from a simple object (such as is imported from YAML).}
\usage{
convert_yaml_to_cdata_spec(obj)
}
\arguments{
\item{obj}{object to convert}
}
\value{
cdata transform specification
}
\description{
Read a cdata record transform from a simple object (such as is imported from YAML).
}
