% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineOldFns.R
\name{define_old_cdata_functions}
\alias{define_old_cdata_functions}
\title{Load old (0.5.1) functions into environment, to help code that has not yet adapted.}
\usage{
define_old_cdata_functions(env = parent.frame())
}
\arguments{
\item{env}{environment to define functions in}
}
\value{
list of mappings performed
}
\description{
Load old (0.5.1) functions into environment, to help code that has not yet adapted.
}
