% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{draw_frame}
\alias{draw_frame}
\title{Render a data.frame in draw_frame form.}
\usage{
draw_frame(x, ..., time_format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{x}{data.frame (atomic types, with at least 1 row and 1 column).}

\item{...}{not used for values, forces later arguments to bind by name.}

\item{time_format}{character, format for "POSIXt" classes.}
}
\value{
chracter
}
\description{
Render a data.frame in draw_frame form.
}
\examples{

tc_name <- "training"
x <- build_frame(
   "measure",                   tc_name, "validation" |
   "minus binary cross entropy",      5, 7            |
   "accuracy",                      0.8, 0.6          )
print(x)
cat(draw_frame(x))

}
\seealso{
\code{\link{build_frame}},  \code{\link{qchar_frame}}
}
