% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{get_transform_details}
\alias{get_transform_details}
\title{Upack details of a cdata record transform.}
\usage{
get_transform_details(x)
}
\arguments{
\item{x}{blocks_to_rowrecs_spec or rowrecs_to_blocks_spec}
}
\value{
detailed fields
}
\description{
Unpack details, especially: generate data frames representing both sides of a transform.
}
\keyword{internal}
