% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{\%pivot\%}
\alias{\%pivot\%}
\title{Convert data shape.}
\usage{
a \%pivot\% b
}
\arguments{
\item{a}{data (data.frame or relop) or cdata_record_spec.}

\item{b}{data (data.frame or relop) or cdata_record_spec.}
}
\value{
blocks_to_rowrecs() result.
}
\description{
If \code{b} is a \code{cdata_record_spec} then call \code{blocks_to_rowrecs()} (multiply or join out to the record spec).
If \code{a} is a \code{cdata_record_spec} then call \code{rowrecs_to_blocks()} (aggregate or project from the record spec).
}
\examples{

d <- wrapr::build_frame(
  "id", "measure", "value" |
  1   , "AUC"    , 0.7     |
  1   , "R2"     , 0.4     |
  2   , "AUC"    , 0.8     |
  2   , "R2"     , 0.5     )

record_spec <- new_record_spec(
  wrapr::qchar_frame(
    measure, value |
    AUC    , "AUC" |
    R2     , "R2"  ),
  recordKeys = "id")

(d2 <- record_spec \%pivot\% d)

# identity
( record_spec \%pivot\% d ) \%pivot\% record_spec

# identity
record_spec \%pivot\% d \%pivot\% record_spec

# identity
record_spec \%pivot\% ( d2 \%pivot\% record_spec )

}
