% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{\%//\%}
\alias{\%//\%}
\title{Divide or aggregate block records into row records.}
\usage{
table \%//\% record_spec
}
\arguments{
\item{table}{data (data.frame or relop).}

\item{record_spec}{a cdata_record_spec.}
}
\value{
blocks_to_rowrecs() result.
}
\description{
Call \code{blocks_to_rowrecs()}.
}
