% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{\%**\%}
\alias{\%**\%}
\title{Multiply/join row records into block records.}
\usage{
table \%**\% transform
}
\arguments{
\item{table}{data (data.frame or relop).}

\item{transform}{a rowrecs_to_blocks_spec.}
}
\value{
rowrecs_to_blocks() result.
}
\description{
Call \code{rowrecs_to_blocks()}.
}
\examples{

d <- wrapr::build_frame(
  "id", "AUC", "R2" |
  1   , 0.7  , 0.4  |
  2   , 0.8  , 0.5  )

transform <- rowrecs_to_blocks_spec(
  wrapr::qchar_frame(
    "measure", "value" |
    "AUC"    , AUC     |
    "R2"     , R2      ),
  recordKeys = "id")

d \%**\% transform

# identity (in structure)
d \%**\% transform \%//\% t(transform)

}
