% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{layout_by.cdata_general_transform_spec}
\alias{layout_by.cdata_general_transform_spec}
\title{Use transform spec to layout data.}
\usage{
\method{layout_by}{cdata_general_transform_spec}(transform, table)
}
\arguments{
\item{transform}{object of class blocks_to_rowrecs_spec.}

\item{table}{data.frame or relop.}
}
\value{
re-arranged data or data reference (relop).
}
\description{
Use transform spec to layout data.
}
