% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowsColumns.R
\name{moveValuesToColumns}
\alias{moveValuesToColumns}
\title{Move values from rows to columns (wrapper for \code{tidyr::spread} or pivot).}
\usage{
moveValuesToColumns(data, columnToTakeKeysFrom, columnToTakeValuesFrom,
  rowKeyColumns, allowNAKeys = FALSE, fill = NA, convert = FALSE,
  drop = TRUE, sep = NULL)
}
\arguments{
\item{data}{data.frame to work with.}

\item{columnToTakeKeysFrom}{character name of column build new column names from.}

\item{columnToTakeValuesFrom}{character name of column to get values from.}

\item{rowKeyColumns}{character array names columns that should be table keys.}

\item{allowNAKeys}{logical if FALSE throw if there are NAs in the key columns.}

\item{fill}{passed to \code{tidyr::spread}}

\item{convert}{passed to \code{tidyr::spread}}

\item{drop}{passed to \code{tidyr::spread}}

\item{sep}{passed to \code{tidyr::spread}}
}
\value{
new data.frame with values moved to columns.
}
\description{
Move values from rows to columns (wrapper for \code{tidyr::spread} or pivot).
}
\examples{

d <- data.frame(meas= c('AUC', 'R2'), val= c(0.6, 0.2))
moveValuesToColumns(d,
                    columnToTakeKeysFrom= 'meas',
                    columnToTakeValuesFrom= 'val',
                    rowKeyColumns= c())

}
