% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowsColsQ.R
\name{moveValuesToColumnsD}
\alias{moveValuesToColumnsD}
\title{Map sets rows to columns (query based).}
\usage{
moveValuesToColumnsD(tallTable, keyColumns, controlTable, ...,
  columnsToCopy = NULL, strict = FALSE, checkNames = TRUE,
  showQuery = FALSE, defaultValue = NULL, dropDups = FALSE,
  env = parent.frame())
}
\arguments{
\item{tallTable}{data,frame containing data to be mapped (in-memory data.frame).}

\item{keyColumns}{character list of column defining row groups}

\item{controlTable}{table specifying mapping (local data frame)}

\item{...}{force later arguments to be by name.}

\item{columnsToCopy}{character list of column names to copy}

\item{strict}{logical, if TRUE check control table contents for uniqueness}

\item{checkNames}{logical, if TRUE check names}

\item{showQuery}{if TRUE print query}

\item{defaultValue}{if not NULL literal to use for non-match values.}

\item{dropDups}{logical if TRUE supress duplicate columns (duplicate determined by name, not content).}

\item{env}{environment to look for "winvector_temp_db_handle" in.}
}
\value{
wide table built by mapping key-grouped tallTable rows to one row per group
}
\description{
Transform data facts from rows into additional columns using SQL
and controlTable.
}
\details{
This is using the theory of "fluid data"n
(\url{https://github.com/WinVector/cdata}), which includes the
principle that each data cell has coordinates independent of the
storage details and storage detail dependent coordinates (usually
row-id, column-id, and group-id) can be re-derived at will (the
other principle is that there may not be "one true preferred data
shape" and many re-shapings of data may be needed to match data to
different algorithms and methods).

The controlTable defines the names of each data element in the two notations:
the notation of the tall table (which is row oriented)
and the notation of the wide table (which is column oriented).
controlTable[ , 1] (the group label) cross colnames(controlTable)
(the column labels) are names of data cells in the long form.
controlTable[ , 2:ncol(controlTable)] (column labels)
are names of data cells in the wide form.
To get behavior similar to tidyr::gather/spread one builds the control table
by running an appropiate query over the data.

Some discussion and examples can be found here:
\url{https://winvector.github.io/FluidData/FluidData.html} and
here \url{https://github.com/WinVector/cdata}.
}
\examples{

# pivot example
d <- data.frame(meas = c('AUC', 'R2'), val = c(0.6, 0.2))

cT <- buildPivotControlTableD(d,
                              columnToTakeKeysFrom= 'meas',
                              columnToTakeValuesFrom= 'val')
moveValuesToColumnsD(d,
                     keyColumns = NULL,
                     controlTable = cT)

}
\seealso{
\code{\link{moveValuesToRowsN}}, \code{\link{buildPivotControlTableD}}
}
