% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowsColumns.R
\name{moveValuesToRows}
\alias{moveValuesToRows}
\title{Move values from columns to rows (wrapper for \code{\link[tidyr]{gather}}, or anti-pivot).}
\usage{
moveValuesToRows(data, nameForNewKeyColumn, nameForNewValueColumn,
  columnsToTakeFrom, ..., nameForNewClassColumn = NULL, na.rm = FALSE,
  convert = FALSE, factor_key = FALSE)
}
\arguments{
\item{data}{data.frame to work with.}

\item{nameForNewKeyColumn}{character name of column to write new keys in.}

\item{nameForNewValueColumn}{character name of column to write new values in.}

\item{columnsToTakeFrom}{character array names of columns to take values from.}

\item{...}{force later argumets to bind by name.}

\item{nameForNewClassColumn}{optional name to land original cell classes to.}

\item{na.rm}{passed to \code{\link[tidyr]{gather}}}

\item{convert}{passed to \code{\link[tidyr]{gather}}}

\item{factor_key}{passed to \code{\link[tidyr]{gather}}}
}
\value{
new data.frame with values moved to rows.
}
\description{
For a tutorial please try \code{vignette('RowsAndColumns', package='cdata')}.
}
\examples{

d <- data.frame(AUC= 0.6, R2= 0.2)
moveValuesToRows(d,
                 nameForNewKeyColumn= 'meas',
                 nameForNewValueColumn= 'val',
                 columnsToTakeFrom= c('AUC', 'R2'))

}
\seealso{
\code{\link{moveValuesToColumns}}
}
