% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{new_record_spec}
\alias{new_record_spec}
\title{Create a record specification.}
\usage{
new_record_spec(controlTable, ..., recordKeys = character(0),
  controlTableKeys = colnames(controlTable)[[1]])
}
\arguments{
\item{controlTable}{an all character data frame or cdata pivot control.}

\item{recordKeys}{vector of columns identifying records.}

\item{controlTableKeys}{vector of keying columns of the controlTable.}
}
\value{
a record specification object
}
\description{
Create a record specification object that holds the pivot control table, spefication of
extra row keys, and control table keys.
}
