% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{qchar_frame}
\alias{qchar_frame}
\title{Build a quoted data.frame}
\usage{
qchar_frame(..., cf_eval_environment = parent.frame())
}
\arguments{
\item{...}{cell names, first infix operator denotes end of header row of column names.}

\item{cf_eval_environment}{environment to evaluate names in.}
}
\value{
character data.frame
}
\description{
A confineint way to build a character data.frame.  Position of
first "+" (or other infix operator) determines number of columns
(all other infic operators are aliases for ",").
Names are treated as character types.
}
\examples{

qchar_frame(
   measure,                      training, validation /
   "minus binary cross entropy", loss,     val_loss   /
   accuracy,                     acc,      val_acc    )

}
\seealso{
\link{char_frame}
}
