% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%.>\%}
\alias{reexports}
\alias{let}
\alias{reexports}
\alias{qc}
\alias{reexports}
\alias{build_frame}
\alias{reexports}
\alias{draw_frame}
\alias{reexports}
\alias{qchar_frame}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{wrapr}{\code{\link[wrapr]{\%.>\%}}, \code{\link[wrapr]{let}}, \code{\link[wrapr]{qc}}, \code{\link[wrapr]{build_frame}}, \code{\link[wrapr]{draw_frame}}, \code{\link[wrapr]{qchar_frame}}}
}}

