% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{char_frame}
\alias{char_frame}
\title{Build character data.frame}
\usage{
char_frame(..., cf_eval_environment = parent.frame())
}
\arguments{
\item{...}{cell names, first infix operator denotes end of header row of column names.}

\item{cf_eval_environment}{environment to evaluate names in.}
}
\value{
character data.frame
}
\description{
A confineint way to build a character data.frame.  Position of
first "+" (or other infix operator) determines number of columns
(all other infix operators are aliases for ",").
Names are de-referenced.
}
\examples{

tc_name <- "training"
char_frame(
   "measure",                   tc_name, "validation" +
   "minus binary cross entropy", "loss", "val_loss"   +
   "accuracy",                    "acc", "val_acc"    )

}
\seealso{
\link{qchar_frame}
}
