% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTransform.R
\name{convert_records}
\alias{convert_records}
\title{General transform from arbitrary record shape to arbitrary record shape.}
\usage{
convert_records(table, incoming_shape, outgoing_shape, ...,
  keyColumns = NULL, columnsToCopy_in = NULL, checkNames = TRUE,
  checkKeys = FALSE, strict = FALSE, use_data_table = FALSE,
  tmp_name_source = wrapr::mk_tmp_name_source("bltrr"),
  temporary = TRUE)
}
\arguments{
\item{table}{data.frame or relop.}

\item{incoming_shape}{data.frame, definition of incoming record shape.}

\item{outgoing_shape}{data.frame, defintion of outgoing record shape.}

\item{...}{force later arguments to bind by name.}

\item{keyColumns}{character vector of column defining incoming row groups}

\item{columnsToCopy_in}{character array of incoming column names to copy.}

\item{checkNames}{logical, if TRUE check names.}

\item{checkKeys}{logical, if TRUE check columnsToCopy form row keys (not a requirement, unless you want to be able to invert the operation).}

\item{strict}{logical, if TRUE check control table name forms.}

\item{use_data_table}{logical if TRUE try to use data.table for the un-pivots.}

\item{tmp_name_source}{a tempNameGenerator from cdata::mk_tmp_name_source()}

\item{temporary}{logical, if TRUE use temporary tables}
}
\value{
processing pipeline or transformed table
}
\description{
General transform from arbitrary record shape to arbitrary record shape.
}
\examples{


incoming_shape <- qchar_frame(
  row,  col1, col2, col3 |
  row1, v11,  v12,  v13  |
  row2, v21,  v22,  v23  |
  row3, v31,  v32,  v33  )


outgoing_shape <- qchar_frame(
  column, row1, row2, row3 |
  col1,   v11,  v21,  v31  |
  col2,   v12,  v22,  v32  |
  col3,   v13,  v23,  v33  )

data <- build_frame(
  'record_id', 'row',  'col1', 'col2', 'col3' |
  1,           'row1',  1,      2,      3      |
  1,           'row2',  4,      5,      6      |
  1,           'row3',  7,      8,      9      |
  2,           'row1',  11,     12,     13      |
  2,           'row2',  14,     15,     16      |
  2,           'row3',  17,     18,     19      )

print(data)

convert_records(
  data,
  keyColumns = 'record_id',
  incoming_shape = incoming_shape,
  outgoing_shape = outgoing_shape)

td <- rquery::local_td(data)

ops <- convert_records(
  td,
  keyColumns = 'record_id',
  incoming_shape = incoming_shape,
  outgoing_shape = outgoing_shape)

cat(format(ops))

if(requireNamespace("rqdatatable", quietly = TRUE)) {
  library("rqdatatable")
  data \%.>\%
    ops \%.>\%
    print(.)
}


}
