% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{\%**\%}
\alias{\%**\%}
\title{Multiply row records into block records.}
\usage{
table \%**\% record_spec
}
\arguments{
\item{table}{data (data.frame or relop).}

\item{record_spec}{a cdata_record_spec.}
}
\value{
rowrecs_to_blocks() result.
}
\description{
Call \code{rowrecs_to_blocks()}.
}
