% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opspec.R
\name{\%**\%}
\alias{\%**\%}
\title{Multiply/join row records into block records.}
\usage{
table \%**\% record_spec
}
\arguments{
\item{table}{data (data.frame or relop).}

\item{record_spec}{a cdata_record_spec.}
}
\value{
rowrecs_to_blocks() result.
}
\description{
Call \code{rowrecs_to_blocks()}.
}
\examples{

d <- wrapr::build_frame(
  "id", "AUC", "R2" |
  1   , 0.7  , 0.4  |
  2   , 0.8  , 0.5  )

record_spec <- new_record_spec(
  wrapr::qchar_frame(
    measure, value |
    AUC    , "AUC" |
    R2     , "R2"  ),
  recordKeys = "id")

d \%**\% record_spec

# identity
d \%**\% record_spec \%//\% record_spec

}
