% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowsColsQ.R
\name{qlook}
\alias{qlook}
\title{Quick look at remote data}
\usage{
qlook(my_db, tableName, displayRows = 10, countRows = TRUE)
}
\arguments{
\item{my_db}{DBI database handle}

\item{tableName}{name of table to look at}

\item{displayRows}{number of rows to sample}

\item{countRows}{logical, if TRUE return row count.}
}
\value{
str-line view of data
}
\description{
Quick look at remote data
}
\examples{

my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DBI::dbWriteTable(my_db,
                  'd',
                  data.frame(AUC = 0.6, R2 = 0.2),
                  overwrite = TRUE,
                  temporary = TRUE)
qlook(my_db, 'd')

}
